Move shim client into subpackage
